Refactor ACP config selector flow
