+++ lua/avante/init.lua
@@ -117,8 +117,12 @@ function H.keymaps()
   vim.keymap.set("n", "<Plug>(AvanteConflictNextConflict)", function() Diff.find_next("ours") end)
   vim.keymap.set("n", "<Plug>(AvanteConflictPrevConflict)", function() Diff.find_prev("ours") end)
   vim.keymap.set("n", "<Plug>(AvanteSelectModel)", function() require("avante.api").select_model() end)
-  vim.keymap.set("n", "<Plug>(AvanteSelectACPModel)", function() vim.cmd("AvanteACPModels") end, { noremap = true })
-  vim.keymap.set("n", "<Plug>(AvanteSelectACPMode)", function() vim.cmd("AvanteACPModes") end, { noremap = true })
+  vim.keymap.set("n", "<Plug>(AvanteSelectACPModel)", function() require("avante.api").select_acp_model() end, {
+    noremap = true,
+  })
+  vim.keymap.set("n", "<Plug>(AvanteSelectACPMode)", function() require("avante.api").select_acp_mode() end, {
+    noremap = true,
+  })
 
   if Config.behaviour.auto_set_keymaps then
     Utils.safe_keymap_set(
@@ -203,13 +207,13 @@ function H.keymaps()
     Utils.safe_keymap_set(
       "n",
       Config.mappings.select_acp_model,
-      function() vim.cmd("AvanteACPModels") end,
+      function() require("avante.api").select_acp_model() end,
       { desc = "avante: select ACP model" }
     )
     Utils.safe_keymap_set(
       "n",
       Config.mappings.select_acp_mode,
-      function() vim.cmd("AvanteACPModes") end,
+      function() require("avante.api").select_acp_mode() end,
       { desc = "avante: select ACP mode" }
     )
